switch to yeslogic-fontconfig-sys
instead of servo-fontconfig

servo-fontconfig statically links a vendored fontconfig library if
fontconfig is not found by pkgconfig. Using a vendored fontconfig
library instead of the system fontconfig library doesn't actually work
well though:
https://github.com/slint-ui/slint/issues/88
Building a vendored copy of fontconfig is also problematic because
fontconfig has a lot of C dependencies, which makes it difficult to
cross compile the vendored copy of fontconfig:

$ pkg-config fontconfig --static --libs
-lfontconfig -lfreetype -lz -lbz2 -lpng16 -lm -lm -lz -lharfbuzz -lm
-lglib-2.0 -lm -lpcre -lsysprof-capture-4 -pthread -lgraphite2
-lbrotlidec -lbrotlicommon -lxml2 -lz -llzma -lm

Instead of using a vendored copy, with
https://github.com/yeslogic/fontconfig-rs/pull/12
yeslogic-fontconfig-sys will have a Cargo feature to dlopen fontconfig
at runtime instead of linking it at build time. This is exposed in
font-kit with the new source-fontconfig-dlopen feature, which is
disabled by default. The feature can be enabled by setting the
RUST_FONTCONFIG_DLOPEN environment variable to avoid needing to
propagate the Cargo feature all the way through downstream
Cargo.toml's. This feature makes it considerably easier to cross
compile by avoiding the need to cross compile fontconfig and all its
dependencies.